Add dynamic `success_redirect` URLs to be generated (#57)
This allows a `proc` to be used as a `success_redirect` value. This is
particularly useful if you need to generated URLs based on the one of
the parameters returned or applied during the callback function.

It is of particular interest in CoderDojo frontend, which needs to send
a user to Zen to log in, but also redirect the user back to the correct
page in CoderDojo frontend afterwards. So the Zen URL will need to have
a dynamic returnTo parameter based on the original returnTo param
supplied to omniauth.

## Considerations

* The proc is executed using `instance_exec` to make sure it is run in
the context of the controller, rather than the context of the
configuration block where it is first defined.
* I've tried to document this as best as possible!