DRAFT: Toggle openid connect scope at runtime
TODO: Add tests if we think this is a sensible approach/change.

The Profile app uses the presence of the `student` scope in the login
request to decide whether to display the student login form (i.e. the
form containing the school code textbox). We want to allow students (as
well as non-students) to login to Experience CS and so need a way of
changing the `scope` accordingly. By default the `scope` is fixed at
Rails initialization time in the `RpiAuth::Engine`[1] but we need to be
able to toggle it at runtime depending on whether user is a student or
not.  By utilising OmniAuth's setup phase[2] we can toggle the `scope`
based on the presence of data in the Rails session. In this case we're
toggling it if the Rails app has set `rpi_auth.login_type` to `student`
in the session.

[1]: https://github.com/RaspberryPiFoundation/rpi-auth/blob/b7771ee120254e4c6f2d90c5025be5714daeb542/lib/rpi_auth/engine.rb#L31
[2]: https://github.com/omniauth/omniauth/wiki/Setup-Phase