Move Authenticatable#attributes out of included block
This is more idiomatic use of an `ActiveSupport::Concern`. Instance
methods defined on the concern module will be available in the class
that includes the concern and the method will also be available in the
ancestor chain for overriding and the use of `super`. The `included`
block is intended for calling class-level "macros" like `attr_accessor`;
not for defining instance methods.