Add explicit require for Controllers::CurrentUser
Previously `RpiAuth::Controllers::CurrentUser` (which is in the `lib`
directory) was only being required somewhat by accident via
`RpiAuth::AuthController` which is required automatically, because it's
in the `app/controllers` directory.

I think a better solution would be to move the `CurrentUser` concern
into the more idiomatic `app/controllers` directory (or perhaps
`app/controllera/concerns`) where it would be automatically loaded and
there would be no need for an expicit `require`. However, that feels
like a bigger piece of work.