Add RpiAuth::Models::AccountTypes#student_account?
I've extracted this method out of experience-cs [1]. This seems like a
more logical home for it and it might be useful in other apps.

There's some related functionality in editor-api for removing the
"student:" prefix from the user's ID which might make sense to extract
into here too [2], especially if we end up needing to do the same thing
in experience-cs which seems likely.

Since you have to explicitly include the new concern into the `User`
model in your app, this should be a non-breaking change.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/blob/837bf97d4aec7ed915f87f9dd9253b87e2dd00b0/app/models/user.rb#L19-L21
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/ecb784299eea43654660111aa509eedc8270346d/app/models/user.rb#L109