Add RpiAuth::Models::Roles#parsed_roles
I've extracted this method out of experience-cs [1]. This seems like a
more logical home for it and we could avoid the duplication in
editor-api [2] and possibly other apps.

Since you have to explicitly include the new concern into the `User`
model in your app, this should be a non-breaking change.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/blob/837bf97d4aec7ed915f87f9dd9253b87e2dd00b0/app/models/user.rb#L15-L17
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/ecb784299eea43654660111aa509eedc8270346d/app/models/user.rb#L61-L63