Improve isolation between model concern specs
Previously we were defining a `DummyUser` class in each of these specs
and including just the concerns that were needed for the examples.
However, this `DummyUser` class is effectively global and so there was a
possibility of cross-contamination. I could have improved matters by
choosing a different class name in each spec. However, I think it's
better to dynamically create a class in each spec.

A side benefit is that I've managed to fix the four
RSpec/SpecFilePathFormat violations, so I've re-generated the rubocop
to-do list again and incorporated the changes into this commit.