Prepare to release v4.2.0
Since this adds new functionality in a backward compatible manner, it
makes sense for it to be a minor version bump from v4.1.1 -> v4.2.0.

I've split #87 into two lines in the CHANGELOG for clarity.

I've only included the version change of them gem itself in the Gemfile
changes; I discarded changes in versions of other gems which just seemed
to be because the dependencies are unconstrained and they are not needed
for the changes in this release.