Refresh access tokens before expiry
We've seen a small number of exceptions in Experience CS[1] which are
caused by a 401 response from Editor API. I'm fairly confident these are
occurring when the access token used to make the request to Editor API
has expired. My hypothesis is that the token is valid at the time the
request comes into Experience CS (and therefore the auto refresh
behaviour isn't triggered) but is invalid by the time we make the
request to Editor API. This change reduces the risk of this happening by
refreshing the token if it expires in the next 60 seconds.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/issues/914