Add optional on_login_success callback
This adds a new `on_login_success` configuration option. If this option
is set to a `Proc` and the login was successful, the `Proc` will be
called in the context of the `RpiAuth::AuthController#callback` action,
i.e. `current_user` will be available. This is intended to allow apps to
record successful logins.

I'm not convinced that using `instance_exec` is the most sensible
approach, but I decided it was better to be consistent with the way
the `success_redirect` option is implemented.

If the `Proc` raises an exception, this will be rescued and the
exception will be logged as a warning in the Rails log. This is so that
the login will still complete successfully, even if there was a problem
recording it.

The immediate motivation for adding this is so we can record successful
*student* logins in `experience-cs` in order to be able to calculate the
"active users" metric described in this issue [1]. Unfortunately while
this information is available in the `profile` app database for regular
users; it's not available for student users.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/issues/727