Return boolean from AccountTypes#student_account?
Previously this method returned either `0` or `nil`. This works fine in
Ruby, because `0` is truth-y, but it causes problems if you want to
persist the value in a Postgres boolean type, because `0` is cast to
`FALSE`. The new implementation always returns either `true` or `false`.

This will mean we can remove this workaround [1] in experience-cs.

[1]: https://github.com/RaspberryPiFoundation/experience-cs/commit/130592b833509a52aac1511bb17320a4d7eadeb4